DDF-4481 Add the IS EMPTY operator in search queries (#4468)

+++ ui/packages/catalog-ui-search/src/main/webapp/component/filter/filter.view.js
@@ -29,6 +29,13 @@ const ValueModel = require('../value/value.js')
 const CQLUtils = require('../../js/CQLUtils.js')
 const properties = require('../../js/properties.js')
 const Common = require('../../js/Common.js')
+import {
+  geometryComparators,
+  dateComparators,
+  stringComparators,
+  numberComparators,
+  booleanComparators,
+} from './comparators'
 
 const generatePropertyJSON = (value, type, comparator) => {
   const propertyJSON = _.extend({}, metacardDefinitions.metacardTypes[type], {
@@ -176,6 +183,7 @@ module.exports = Marionette.LayoutView.extend({
       CONTAINS: 'ILIKE',
       MATCHCASE: 'LIKE',
       EQUALS: '=',
+      'IS EMPTY': 'IS NULL',
       '>': '>',
       '<': '<',
       '=': '=',
@@ -205,9 +213,13 @@ module.exports = Marionette.LayoutView.extend({
     }
   },
   updateTypeDropdown: function() {
-    this.filterAttribute.currentView.model.set('value', [
-      this.model.get('type'),
-    ])
+    const attribute = this.model.get('type')
+    if (attribute === 'anyGeo') {
+      this.model.set('comparator', [geometryComparators[1]])
+    } else if (attribute === 'anyText') {
+      this.model.set('comparator', [stringComparators[1]])
+    }
+    this.filterAttribute.currentView.model.set('value', [attribute])
   },
   handleAttributeUpdate: function() {
     this.model.set(
@@ -230,23 +242,19 @@ module.exports = Marionette.LayoutView.extend({
     var currentComparator = this.model.get('comparator')
     switch (propertyJSON.type) {
       case 'LOCATION':
-        if (['INTERSECTS'].indexOf(currentComparator) === -1) {
+        if (geometryComparators.indexOf(currentComparator) === -1) {
           this.model.set('comparator', 'INTERSECTS')
         }
-        this.toggleLocationClass(true)
+        this.toggleLocationClass(currentComparator === 'IS EMPTY')
         break
       case 'DATE':
-        if (
-          ['BEFORE', 'AFTER', 'RELATIVE', 'BETWEEN'].indexOf(
-            currentComparator
-          ) === -1
-        ) {
+        if (dateComparators.indexOf(currentComparator) === -1) {
           this.model.set('comparator', 'BEFORE')
         }
         this.toggleDateClass(true)
         break
       case 'BOOLEAN':
-        if (['='].indexOf(currentComparator) === -1) {
+        if (booleanComparators.indexOf(currentComparator) === -1) {
           this.model.set('comparator', '=')
         }
         break
@@ -255,15 +263,12 @@ module.exports = Marionette.LayoutView.extend({
       case 'FLOAT':
       case 'INTEGER':
       case 'SHORT':
-        if (['>', '<', '=', '>=', '<='].indexOf(currentComparator) === -1) {
+        if (numberComparators.indexOf(currentComparator) === -1) {
           this.model.set('comparator', '>')
         }
         break
       default:
-        if (
-          ['CONTAINS', 'MATCHCASE', '=', 'NEAR'].indexOf(currentComparator) ===
-          -1
-        ) {
+        if (stringComparators.indexOf(currentComparator) === -1) {
           this.model.set('comparator', 'CONTAINS')
         }
         break
@@ -285,13 +290,8 @@ module.exports = Marionette.LayoutView.extend({
     let value = Common.duplicate(this.model.get('value'))
     const currentComparator = this.model.get('comparator')
     value = this.transformValue(value, currentComparator)
-    const propertyJSON = generatePropertyJSON(
-      value,
-      this.model.get('type'),
-      currentComparator
-    )
-    let type = this.model.get('type')
-
+    const type = this.model.get('type')
+    const propertyJSON = generatePropertyJSON(value, type, currentComparator)
     if (this.options.suggester && propertyJSON.enum === undefined) {
       this.options.suggester(propertyJSON).then(suggestions => {
         if (suggestions.length > 0) {
@@ -322,6 +322,10 @@ module.exports = Marionette.LayoutView.extend({
     } else {
       this.turnOffEditing()
     }
+    this.$el.toggleClass(
+      'is-empty',
+      this.model.get('comparator') === 'IS EMPTY'
+    )
     this.setDefaultComparator(propertyJSON)
   },
   getValue: function() {
@@ -333,9 +337,10 @@ module.exports = Marionette.LayoutView.extend({
     return text
   },
   getFilters: function() {
-    var property = this.model.get('type')
-    var comparator = this.model.get('comparator')
-    var value = this.filterInput.currentView.model.getValue()[0]
+    const property = this.model.get('type')
+    const comparator = this.model.get('comparator')
+    const value = this.filterInput.currentView.model.getValue()[0]
+    const type = this.comparatorToCQL()[comparator]
 
     if (comparator === 'NEAR') {
       return CQLUtils.generateFilterForFilterFunction('proximity', [
@@ -343,12 +348,16 @@ module.exports = Marionette.LayoutView.extend({
         value.distance,
         value.value,
       ])
+    } else if (comparator === 'IS EMPTY') {
+      return CQLUtils.generateIsEmptyFilter(property)
     }
 
-    var type = this.comparatorToCQL()[comparator]
     return CQLUtils.generateFilter(type, property, value)
   },
   deleteInvalidFilters: function() {
+    if (this.model.attributes.comparator === 'IS EMPTY') {
+      return
+    }
     if (!this.filterInput.currentView.isValid()) {
       this.delete()
     }
